% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_winputall_data.R
\docType{data}
\name{my_winputall_data}
\alias{my_winputall_data}
\title{Example data}
\format{
A data frame with 175 rows and 9 variables:
\describe{
\item{id}{individual index}
\item{year}{time index}
\item{tx}{Total input uses}
\item{s_crop1}{acreage of crop number 1}
\item{s_crop2}{acreage of crop number 2}
\item{s_crop3}{acreage of crop number 3}
\item{k}{Capital}
\item{tempmoy}{Temperature}
\item{precip}{Annual precipitation}
...
}
}
\usage{
my_winputall_data
}
\description{
An unbalanced panel data
}
\keyword{datasets}
