% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_otherfunctions.R
\name{rp_model_fun}
\alias{rp_model_fun}
\title{Random Parameters Transformation}
\usage{
rp_model_fun(beta, opt)
}
\arguments{
\item{beta}{vector of random parameters. beta follows multivariate normal distribution.}

\item{opt}{a list of control parameters. See rpinpall.}
}
\value{
the transformed beta
}
\description{
rp.model.fun is used to tranform the random parameters.
}
