% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winputall-package.R
\docType{package}
\name{winputall-package}
\alias{winputall-package}
\alias{winputall}
\title{The 'winputall' package.}
\description{
Using a time-varying random parameters model, this package allows
allocating variable input costs among crops produced by farmers based on panel
data including information on input expenditure aggregated at the farm level
and acreage shares. It also considers in fairly way the weighting data and
can allow integrating time-varying and time-constant control variables.
}
\references{
Koutchade Obafèmi Philippe, Fabienne Femenia, Alain Carpentier (2024).Variable Input Allocation Among Crops: A Time-Varying Random Parameters Approach https://hal.science/hal-04318163
}
\author{
\strong{Maintainer}: Obafèmi Philippe Koutchade \email{obafemi-philippe.koutchade@inrae.fr} (\href{https://orcid.org/0000-0001-7327-3139}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Fabienne Femenia \email{fabienne.femenia@inrae.fr}
  \item Alain Carpentier \email{alain.carpentier@inrae.fr}
}

}
